0.0.4 Branches now Possible - Redux
+ Improved starting branch positioning
+ Added ability to create a simple branch

This update adds the ability to create branches, still not perfect, can
completely break the branches if not careful. However the ability for
branches to grow and develop is indeed there.

This is technically a second commit of the same commit, except that I'm
making sure it sends out to social media correctly.

Video :

https://drive.google.com/file/d/0B0QGg7T3tU2sQTBna0szV1FmUGs/view?usp=sharing